Added notification functions and error handling.

+++ scripts/favorites.mjs
@@ -5,15 +5,22 @@ export function addToFavorites(cityData){
         localStorage.setItem('favorites', JSON.stringify(favorites));
     }
     else{
-        console.log('City is already in favorites');
+        throw new Error('City already in favorites');
     }
+    return "City added to favorites. ";
 }
 
 export function removeFromFavorites(cityData){
     let favorites = JSON.parse(localStorage.getItem('favorites')) || [];
-    favorites = favorites.filter(fav => !(fav.id === cityData.id));
-    localStorage.setItem('favorites', JSON.stringify(favorites));
-    console.log('City removed from favorites');
+    const exists = favorites.some(fav => fav.id === cityData.id);
+    if (!exists){
+        throw new Error('City not in favorites');
+    }
+    else{
+        favorites = favorites.filter(fav => !(fav.id === cityData.id));
+        localStorage.setItem('favorites', JSON.stringify(favorites));
+        return "City removed from favorites. ";
+    }
 }
 
 export function getFavorites(){
